Completed function for checking one island integrity
Function returns true if island complete and false if not. Cases in which the island has more tiles than needed are not handled yet. They will unlikely be a problem as it will be easy to verify for the final algorithm that an island is already complete and not add a tile to it.