Added two functions
Card: https://github.com/monblu/nurikabe/projects/1#card-23009999
Added a function to add tiles to an island if and only if it isn't complete and is surrounded by the Wall, everywhere except in one spot. (It keeps relaunching itself until the conditions are not met).
The second triggers the first functions on all islands. Made in two days. First, the function was made to be executed directly on all islands . But, it was triggered only once per island. I wanted it to repeat itself until its conditions are not met. It was easier to make a recursive function that only treats one island and make a second one to launch it on every island. 
Note: there is a "quite deprecated" function isInt. isIntTwo, a better successor is more commonly used. Should replace isInt by isIntTwo.
Note: I thought about one function that would call different functions on all islands. So, instead of ,almost each time, having the main function and the "iterator", having one iterator that would accept a function name as argument and trigger it for each island. Can be tricky.