IMPORTANT update
I find that at least for now, the main big function doesn't need to be recursive. It isn't even a function now but a simple while loop. It's based on  class "state":
```python
class state:
    def __init__(self, table):
        self.table = table
        self.impossibleMoves = []
```
The loop goes through the tree of possible island layouts, executing logic moves and doing suppositions when these run out.
This code is tested for the following table only, I worked several weeks to arrive at this state and would like to push it now as it works.
```python
[[0,0,0,0,0],
             [2,0,0,2,0],
             [0,0,0,0,0],
             [3,0,1,0,0],
             [0,0,0,0,1]]
```
Notes: 
* The "recursion dephth" (just a var that is incremented with each loop execution) reaches to 26 for this table. 
* The algorithm keeps only one "branch" of the all-possibilities-tree and thus should not have problems with memory overflow.
* Potential new problems will be these concerning bigger islands than 2. 
* Some (actually maybe one) functions need to be rewritten, they didn't produce unwanted behaviour for this table but might in the future.